move FigureDataManager.cs to HabboHotel folder
