--- Core/FigureData/Types/Part.cs
@@ -1,19 +1,0 @@
-﻿namespace Plus.Core.FigureData.Types;
-
-public class Part
-{
-    public Part(int id, SetType setType, bool colorable, int index, int colorIndex)
-    {
-        Id = id;
-        SetType = setType;
-        Colorable = colorable;
-        Index = index;
-        ColorIndex = colorIndex;
-    }
-
-    public int Id { get; set; }
-    public SetType SetType { get; set; }
-    public bool Colorable { get; set; }
-    public int Index { get; set; }
-    public int ColorIndex { get; set; }
-}
